Add ITargetFeature
